8918f03 Make it possible to use vtknetcdf4 with external hdf5.
05972ce Move netcdf_cxx option to the advanced option set.
12f985b Merge branch 'qt-wrap-hierarchy-rel' into release
3d65498 Merge branch 'boost-api-change-rel' into release
528adf1 Merge branch 'charts-top-right-axis-bug-rel' into release
e3bc66f Merge branch 'chemistry-python-wrapping-rel' into release
771f45b Merge branch 'charts-valgrind-rel' into release
f5ca601 BUG: Fixed a bug in scrolling of top and right axes.
87e0c1e COMP: API change in Boost Graph Library.
7aaf1d8 BUG: Fix the wrapping of QVTKInteractor.
be52026 BUG: Fix bug 12978 - python chemistry CMake variable.
34298ae Increment version to VTK 5.10.0
e46f3ec KWSys Nightly Date Stamp
05b9f40 Merge topic 'pi-precision' into master
9bcb5ff COMP: Fixed warnings about unitialized memory.
8bd0d55 Merge topic 'vtkImplicitPlaneWidget2' into master
...

+++ Chemistry/vtkCMLMoleculeReader.cxx
@@ -196,7 +196,7 @@ void vtkCMLParser::NewAtom(const char **attr)
   vtkAtom atom = this->Target->AppendAtom();
   int attrInd = 0;
   unsigned short atomicNum = 0;
-  double pos[3];
+  float pos[3];
   const char * id;
   while (const char * cur = attr[attrInd])
     {
